Handle gracefully when reaching EOF or error in stdin read
If stdin is closed (e.g. by using C-D shortcut) or there was an other
error reading from it, there is no point in proceeding with the command
processing.

If stdin is indeed closed, we won't be able to read from it again and we
should simply exit the shell.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>