Finish waiting for child process status after it's terminated
After waitid() exists with a value of 0 and we verify that it hasn't
stopped, all we're left to do is exit the polling loop.

Before this patch the waiting loop just started another iteration
waiting for a process which is already ripped which caused it to print
an error message.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>