Prevent modifying child process id and group id
After forking a child process, the (smash) shell made it obtain the same
process id and group id as the main shell.
This action prevented some program from running and switched immediately
into 'stop' state (e.g. programs like bash/python3/vim).

When doing some experiments with bash, it seems like it doesn't apply
enforce spawned processes to have the same process id (to be honest, it
is somewhat weird this is even possible).

This commit removes this behavior from smash shell for now. Might be
returned if a justification for it is found.

Signed-off-by: Shay Agroskin <shayagr@amazon.com>